Upgrade clj-http, add support for input stream bodies in .clj-http
Had to do a bit of tricks to make an InputStream object that can
serialize itself without consuming itself, by extending the built-in
java.io.ByteArrayInputStream.

Added a new option to the specs called :return-transformer that lets
you tweak the return value of the recording thing before it gets
serialized.