vcr-clj.core/with-cassette supports :arg-transformer option
I added a new `:arg-transformer` option that is analogous to the existing
`:return-transformer` option. During both recording and playback, the original
arguments are intercepted and replaced by `(apply return-transformer args)`,
and these transformed arguments are passed to `arg-key-fn`, `recordable?`, and
the recorded function.

This feature is intended to enable serialization of HTTP request bodies for
https://github.com/gfredericks/vcr-clj/issues/17