Switch cassette extension from .clj to .edn; keep supporting .clj on reads
I think .edn was always a more appropriate extension. This is
especially convenient when using file extensions in a repo to
determine what files have code in them.

If a file with a .clj extension exists, it will still be read for
backwards compatibility, but after deleting and rerecording, .edn will
be used.